% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jags_check.R
\name{jags_check_pp}
\alias{jags_check_pp}
\title{Perform Posterior Predictive Checks}
\usage{
jags_check_pp(.recipe, .model, checks, by_iter = FALSE)
}
\arguments{
\item{.recipe}{A recipe object}

\item{.model}{The output of \code{\link{jags_model_run}}}

\item{checks}{A character string (or vector) of statistical summaries such as c("min", "max", "mean", "sd").}

\item{all_iters}{A logical indicating whether the statistic summaries should be displayed per iteration. FALSE by default.}
}
\value{
Returns a Bayesian P Value for each posterior predictive check.
}
\description{
Perform Posterior Predictive Checks
}
