% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jags_check.R
\name{jags_check_residuals}
\alias{jags_check_residuals}
\title{Get residuals}
\usage{
jags_check_residuals(.recipe, .model)
}
\arguments{
\item{.recipe}{A recipe object}

\item{.model}{The output of \code{\link{jags_model_run}}}
}
\value{
A data frame containing two columns \code{.pred}, which are the estimated
predictor values and \code{.resid}, the difference between the true data and the
predicted data.
}
\description{
Get residuals
}
\examples{
# coming soon
}
