% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jags_data.R
\name{jags_data_num_obs}
\alias{jags_data_num_obs}
\title{Get numbers of observations}
\usage{
jags_data_num_obs(.recipe)
}
\arguments{
\item{.recipe}{A recipe object.}
}
\value{
The number of observations
}
\description{
Get numbers of observations
}
\examples{
# coming soon
}
