% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jags_data.R
\name{jags_data_num_preds}
\alias{jags_data_num_preds}
\title{Get numbers of predictors}
\usage{
jags_data_num_preds(.pred_matrix)
}
\arguments{
\item{.pred_matrix}{The object returned by \code{\link{jags_data_predictors}}.}
}
\value{

}
\description{
Get numbers of predictors
}
