% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jags_data.R
\name{jags_data_predictors}
\alias{jags_data_predictors}
\title{Get the matrix of predictors}
\usage{
jags_data_predictors(.recipe)
}
\arguments{
\item{.recipe}{A recipe object}
}
\value{
A matrix of predictors for the model
}
\description{
Get the matrix of predictors
}
