% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jags_data.R
\name{jags_data_response}
\alias{jags_data_response}
\title{Get the response vector}
\usage{
jags_data_response(.recipe)
}
\arguments{
\item{.recipe}{A recipe object}
}
\value{

}
\description{
Get the response vector
}
\examples{
# coming soon
}
