% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jags_model.R
\name{jags_model_run}
\alias{jags_model_run}
\title{Run JAGS models}
\usage{
jags_model_run(.jags_model, num_iter, variable_names, checks = NULL, ...)
}
\arguments{
\item{.jags_model}{The output of \code{\link{jags_model_define}}}

\item{num_iter}{Number of iterations per chain.}

\item{variable_names}{The variables to save from JAGS output.}

\item{...}{Additional arguments for \link[rjags2]{coda.samples}}
}
\value{

}
\description{
Run JAGS models
}
