% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jags_data.R
\name{jags_data_prep}
\alias{jags_data_prep}
\title{Prep Matrices}
\usage{
jags_data_prep(.recipe, family)
}
\arguments{
\item{.recipe}{A recipe object}

\item{family}{The type of model to perform (e.g. "lm").}
}
\value{
A list to be passed into \code{jags.model}
}
\description{
Prep Matrices
}
