% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jags_model.R
\name{jags_model_lm}
\alias{jags_model_lm}
\title{Define JAGS model}
\usage{
jags_model_lm(jags_data, num_chains, ...)
}
\arguments{
\item{jags_data}{The output of \code{\link{jags_data_prep}}}

\item{num_chains}{How many chains for the MCMC.}

\item{...}{Additional arguments for \link[rjags2]{jags.model}}
}
\value{

}
\description{
Define JAGS model
}
